ivar2: Use connection.uri
This change makes it possible to specify tcpv6 and tls in connection.
Not that I've had some problems with binding laddr, so I removed the
example for now.

This change also improves the hostmask, instead of relying on user to
specify correct hostmask we parse the incoming hostmask from the server
on JOINs.

Conflicts:
	ivar2.lua

Former-commit-id: 9a07905ea45338a2e6389847b08b55063a681cc4